+++ example/lib/example_page_items.dart
@@ -30,9 +30,10 @@ import 'pages/radio_page.dart';
 import 'pages/search_field_page.dart';
 import 'pages/section_page.dart';
 import 'pages/selectable_container_page.dart';
+import 'pages/split_button_page.dart';
 import 'pages/switch_page.dart';
 import 'pages/tab_bar_page.dart';
-import 'pages/theme_page/theme_page.dart';
+import 'pages/theme_page/home.dart';
 import 'pages/tile_page.dart';
 import 'pages/window_controls_page.dart';
 
@@ -326,12 +327,11 @@ final examplePageItems = <PageItem>[
   ),
   PageItem(
     title: 'YaruIcons',
-    titleBuilder: createIconsPageAppBarTitle,
-    actionsBuilder: createIconsPageAppBarActions,
-    floatingActionButtonBuilder: createIconsPageFloatingActionButton,
-    pageBuilder: (context) {
-      return const IconsPage();
-    },
+    titleBuilder: (context) => const IconsPageAppBarTitle(),
+    actionsBuilder: (context) => [const IconsSearchIcon()],
+    floatingActionButtonBuilder: (context) =>
+        const IconsPageFloatingActionButton(),
+    pageBuilder: (context) => const IconsPage(),
     iconBuilder: (context, selected) => selected
         ? const Icon(YaruIcons.placeholder_icon_filled)
         : const Icon(YaruIcons.placeholder_icon),
@@ -348,9 +348,7 @@ final examplePageItems = <PageItem>[
   ),
   PageItem(
     title: 'Material Components, using Yaru Material Themes',
-    pageBuilder: (context) {
-      return const ThemePage();
-    },
+    pageBuilder: (context) => const MaterialThemeHomePage(),
     iconBuilder: (context, selected) => selected
         ? const Icon(YaruIcons.colors_filled)
         : const Icon(YaruIcons.colors),
@@ -380,5 +378,14 @@ final examplePageItems = <PageItem>[
       snippetUrl:
           'https://raw.githubusercontent.com/ubuntu/yaru.dart/main/example/lib/pages/border_container_page.dart',
     ),
+  ),
+  PageItem(
+    title: 'YaruSplitButton',
+    floatingActionButtonBuilder: (_) => const CodeSnippedButton(
+      snippetUrl:
+          'https://raw.githubusercontent.com/ubuntu/yaru.dart/main/example/lib/pages/split_button_page.dart',
+    ),
+    pageBuilder: (context) => const SplitButtonPage(),
+    iconBuilder: (context, selected) => const Icon(YaruIcons.pan_down),
   ),
 ].sortedBy((page) => page.title);